Refactor tests.c to include missing test functions

+++ libft/ft_substr.c
@@ -6,7 +6,7 @@
 /*   By: sfarren <sfarren@student.42malaga.com>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/23 15:53:55 by sfarren           #+#    #+#             */
-/*   Updated: 2024/05/31 17:00:27 by sfarren          ###   ########.fr       */
+/*   Updated: 2024/06/01 19:22:43 by sfarren          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,19 +20,19 @@
  * @param start The start index of the substring in the string 's'.
  * @param len The maximum length of the substring.
  * @return The substring if successful, NULL if the allocation fails.
- 
+
 Check if the start index is greater than the length of the string or not.
 Check if the start plus the len is greater than the length of the whole string.
-Allocate enough memory for the substring, 
+Allocate enough memory for the substring,
 Copy from s[start] until we reach len characters copied into our new string.
 Finally we can return the substring.
- 
+
  */
 char	*ft_substr(char const *s, unsigned int start, size_t len)
 {
-	char	*substr;
-	int		i;
-	int		j;
+	char		*substr;
+	size_t		i;
+	size_t		j;
 
 	if (!s)
 		return (0);
